requirements.txt
google-genai==0.3.0
notion-client==2.2.1
python-telegram-bot==21.0.1